requirements-cpu.txt
# FILE: requirements-cpu.txt
# 终极版：一个自给自足、版本精确锁定的本地CPU测试依赖文件。

# --- 指定额外的包索引地址 ---
# 告诉uv/pip，要去PyTorch的CPU官方地址寻找包
--extra-index-url https://download.pytorch.org/whl/cpu

# --- 核心深度学习框架 (CPU版本) ---
torch==2.3.0
torchvision==0.18.0
numpy<2.0

# --- Hugging Face 生态系统 (核心) ---
transformers==4.41.2
datasets==2.19.0
accelerate==0.30.1
tokenizers==0.19.1
# [新增] TRL (Transformer Reinforcement Learning) 是SFT和DPO训练器必需的核心库
trl==0.8.6
# [新增] PEFT for parameter-efficient fine-tuning techniques like LoRA.
peft==0.10.0
# [新增] Quantization library, also needed for some trainer functionalities.
bitsandbytes==0.43.1

# --- 数据处理与工具库 ---
sentencepiece==0.2.0
pyyaml==6.0.1
Pillow==10.4.0
tqdm==4.66.4
wandb==0.17.0

# --- 代码质量工具 ---
ruff==0.4.4